Organize pages into a 'pages' folder
